feat: initial file system repository impl
This commit introduces an initial implementation of a `Repository` that
reads/writes to the file system using the `RepositoryFileV1` schema.

Presently the implementation reads the file at application start,
creating it if needed. Any records are then written to the file at
application termination. New records are held in memory in the interim.
There is a risk of data loss if the application crashes.